+++ pkgs/io/test/copy_path_test.dart
@@ -5,6 +5,8 @@
 @TestOn('vm')
 library;
 
+import 'dart:io';
+
 import 'package:io/io.dart';
 import 'package:path/path.dart' as p;
 import 'package:test/test.dart';
@@ -32,6 +34,36 @@ void main() {
       ),
       throwsArgumentError,
     );
+  });
+
+  test('links are preserved in async copy', () async {
+    await _create();
+    const linkTarget = 'link_target';
+    final targetPath = p.join(d.sandbox, linkTarget);
+    const linkSource = 'link_source';
+    await d.dir(linkTarget).create();
+    await Link(p.join(d.sandbox, _parentDir, linkSource)).create(targetPath);
+
+    await copyPath(p.join(d.sandbox, _parentDir), p.join(d.sandbox, _copyDir));
+
+    final expectedLink = Link(p.join(d.sandbox, _copyDir, linkSource));
+    expect(await expectedLink.exists(), isTrue);
+    expect(await expectedLink.target(), targetPath);
+  });
+
+  test('links are preserved in sync copy', () async {
+    await _create();
+    const linkTarget = 'link_target';
+    final targetPath = p.join(d.sandbox, linkTarget);
+    const linkSource = 'link_source';
+    await d.dir(linkTarget).create();
+    await Link(p.join(d.sandbox, _parentDir, linkSource)).create(targetPath);
+
+    copyPathSync(p.join(d.sandbox, _parentDir), p.join(d.sandbox, _copyDir));
+
+    final expectedLink = Link(p.join(d.sandbox, _copyDir, linkSource));
+    expect(await expectedLink.exists(), isTrue);
+    expect(await expectedLink.target(), targetPath);
   });
 }